requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
PyJWT==2.3.0
python-decouple==3.6
pydantic[email]